REF: restructure files
